refactor: Update Combo and Order entities
This commit updates the Combo and Order entities. In the Combo entity, the id property has been changed from a string to a number or null. The constructor now accepts an id parameter. In the Order entity, new properties have been added: customer, statusPayment, status, and createdAt. Getter and setter methods have been added for these properties. These changes enhance the structure and functionality of the entities.