feat: Add OrderDto and ProductDto classes
This commit adds the `OrderDto` and `ProductDto` classes to handle the data transfer objects for creating orders and products respectively. These classes define the structure of the data that will be sent in the request body when creating orders and products. This is necessary to ensure that the data is properly validated and processed by the server.