feat: Add identifyCustomer endpoint to CustomerController
This commit adds the `identifyCustomer` endpoint to the `CustomerController`. The endpoint allows clients to identify a customer by their CPF (Brazilian individual taxpayer registry number). If the customer is found, their information is returned in the response body. If the customer is not found, a 404 error is returned with an appropriate error message.