chore: Update tsconfig.json and package.json
This commit updates the `tsconfig.json` file to remove the `tests` directory from the `include` property and removes the trailing comma in the `exclude` property.

It also updates the `package.json` file to update the `eslint` package to version 8.57.0, add the `eslint-config-prettier` package, and add the `eslint-plugin-prettier` and `eslint-plugin-simple-import-sort` packages.

These changes improve the configuration and dependencies of the project.