feat: Add listOrders endpoint to OrderController
This commit adds the `listOrders` endpoint to the `OrderController`. The endpoint allows clients to retrieve a list of all orders. The list of orders is returned in the response body as an array of order objects. This feature enhances the functionality of the application by providing the ability to view all orders in a convenient manner.