refactor: Add paymentGatewayId field to Order model and update related code
This commit adds the `paymentGatewayId` field to the `Order` model in the Prisma schema and updates the corresponding code in the `Order` entity, `OrderUseCase`, `OrderRepository`, and `MercadoPagoPaymentGateway` files. The `paymentGatewayId` field allows for storing the ID of the payment gateway used for the order. This enhancement improves the payment processing functionality of the application.